Initial import of the Python codecs for RISCOS alphabets.
This is the initial version of the Python codecs for RISC OS alphabets.
Found to be necessary, because otherwise trying to print out characters
like 0x80 will behave oddly in some terminals, when actually it has
a completely different meaning on RISC OS - you probably want it to show
a Euro.