Add support for the errors handling in the decode interface.
I'm using the errors='replace' in the place where I used the rocodecs
package, and found that these weren't implemented. I could not understand
how the built in codecs lookup for different error types were meant to
be used, so I've just implemented separate helper methods for the lookups
of characters in the tables.

I have vague concerns about the performance, but having something working
is more important at this point.